rename headers to have unique names
