+++ src/lib/Dirac/mderiv.cu
@@ -20,7 +20,7 @@
 #include "cuda.h"
 #include <cuComplex.h>
 #include <stdio.h>
-#include "GPUtune.h"
+#include "Dirac_GPUtune.h"
 
 /* enable this for checking for kernel failure */
 //#define CUDA_DBG
